rename stencil implementations so non-collective one is default. Add README.

+++ GRAPPA/Stencil/stencil.cpp
@@ -345,7 +345,6 @@ int main(int argc, char * argv[]) {
 		    writeXF( &bottom_halo[kk], val);
 		  } );
 	      }
-	      // FullEmpty<Bool>
 
 	  if (my_IDy > 0 && readFE( CTS_bottom))
 	    for (kk=0,j=jstart; j<=jstart+RADIUS-1; j++)
@@ -426,21 +425,7 @@ int main(int argc, char * argv[]) {
 	  
 	} // end of iterations                                                   */
       } );
-    //LOG(INFO) << "on_all_cores complete. waiting to finish";
       } );       
-
-    //LOG(INFO) << "iterations complete";
-
-
-
-
-	    // // if last data element is empty, go ahead and right through
-	    // Grappa::delegate::call<async>( top_nbr, [=] {
-	    // 	// if fullbit isEmpty, write otherwise suspend
-	    //   }
-
-
-
     
  
     symmetric DTYPE local_norm;
@@ -471,7 +456,7 @@ int main(int argc, char * argv[]) {
       // plus one flop for the update of the input of the array       
       int stencil_size = 4*RADIUS+1;
       double flops = (DTYPE) (2*stencil_size+1) * f_active_points;
-      double avgtime = iter_time/iterations;
+      double avgtime = iter_time/(double)iterations;
       std::cout << "Solution validates"<<std::endl;
       std::cout << "Rate (MFlops/s): " << 1.0E-06*flops/avgtime<<
         "  Avg time (s): "<<avgtime<<std::endl;
